+++ controllers/UserController.js
@@ -1,0 +1,116 @@
+import UserModel from '../models/User.js';
+import jwt from 'jsonwebtoken';
+import bcrypt from 'bcrypt';
+import { validationResult } from 'express-validator';
+
+// Registration
+export const register = async (req, res) => {
+  try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json(errors.array());
+    }
+
+    const password = req.body.password;
+    const salt = await bcrypt.genSalt(10);
+    const hash = await bcrypt.hash(password, salt); //шифруєм пароль
+
+    const doc = new UserModel({
+      email: req.body.email,
+      passwordHash: hash,
+      fullName: req.body.fullName,
+      avatarUrl: req.body.avatarUrl,
+    });
+
+    // генеруємо токен з id користувача та ставимо час його дії 14 днів
+    const user = await doc.save();
+    const token = jwt.sign(
+      {
+        _id: user._id,
+      },
+      'oskar5647',
+      {
+        expiresIn: '14d',
+      }
+    );
+
+    const { passwordHash, ...userData } = user._doc;
+
+    res.json({
+      ...userData,
+      token,
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не вдалось зареєструватись',
+    });
+  }
+};
+
+// Login
+export const login = async (req, res) => {
+  try {
+    const user = await UserModel.findOne({ email: req.body.email });
+
+    if (!user) {
+      return res.status(404).json({
+        message: 'Не вірний логін',
+      });
+    }
+
+    const isValidPass = await bcrypt.compare(
+      req.body.password,
+      user._doc.passwordHash
+    );
+
+    if (!isValidPass) {
+      return res.status(404).json({
+        message: 'Не вірний логін або пароль',
+      });
+    }
+
+    const token = jwt.sign(
+      {
+        _id: user._id,
+      },
+      'oskar5647',
+      {
+        expiresIn: '14d',
+      }
+    );
+
+    const { passwordHash, ...userData } = user._doc;
+
+    res.json({
+      ...userData,
+      token,
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Не вдалось авторизуватись',
+    });
+  }
+};
+
+// Get me
+export const getMe = async (req, res) => {
+  try {
+    const user = await UserModel.findById(req.userId);
+
+    if (!user) {
+      return res.status(404).json({
+        message: 'Користувач не знайдений',
+      });
+    }
+    const { passwordHash, ...userData } = user._doc;
+
+    res.json(userData);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: 'Нема доступу',
+    });
+  }
+};
