+++ index.js
@@ -1,12 +1,8 @@
 import express from 'express';
-import jwt from 'jsonwebtoken';
-import bcrypt from 'bcrypt';
 import mongoose from 'mongoose';
-import { registerValidation } from './validations/auth.js';
-import { validationResult } from 'express-validator';
-import UserModel from './models/User.js';
 import checkAuth from './utils/checkAuth.js';
-import User from './models/User.js';
+import { registerValidation } from './validations/auth.js';
+import * as UserController from './controllers/UserController.js';
 
 // Mongodb
 mongoose
@@ -23,116 +19,11 @@ const app = express();
 app.use(express.json());
 
 // Authorisation
-app.post('/auth/login', async (req, res) => {
-  try {
-    const user = await UserModel.findOne({ email: req.body.email });
-
-    if (!user) {
-      return res.status(404).json({
-        message: 'Не вірний логін',
-      });
-    }
-
-    const isValidPass = await bcrypt.compare(
-      req.body.password,
-      user._doc.passwordHash
-    );
-
-    if (!isValidPass) {
-      return res.status(404).json({
-        message: 'Не вірний логін або пароль',
-      });
-    }
-
-    const token = jwt.sign(
-      {
-        _id: user._id,
-      },
-      'oskar5647',
-      {
-        expiresIn: '14d',
-      }
-    );
-
-    const { passwordHash, ...userData } = user._doc;
-
-    res.json({
-      ...userData,
-      token,
-    });
-  } catch (err) {
-    console.log(err);
-    res.status(500).json({
-      message: 'Не вдалось авторизуватись',
-    });
-  }
-});
-
+app.post('/auth/login', UserController.login);
 // Registration
-app.post('/auth/register', registerValidation, async (req, res) => {
-  try {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json(errors.array());
-    }
-
-    const password = req.body.password;
-    const salt = await bcrypt.genSalt(10);
-    const hash = await bcrypt.hash(password, salt); //шифруєм пароль
-
-    const doc = new UserModel({
-      email: req.body.email,
-      passwordHash: hash,
-      fullName: req.body.fullName,
-      avatarUrl: req.body.avatarUrl,
-    });
-
-    // генеруємо токен з id користувача та ставимо час його дії 14 днів
-    const user = await doc.save();
-    const token = jwt.sign(
-      {
-        _id: user._id,
-      },
-      'oskar5647',
-      {
-        expiresIn: '14d',
-      }
-    );
-
-    const { passwordHash, ...userData } = user._doc;
-
-    res.json({
-      ...userData,
-      token,
-    });
-  } catch (err) {
-    console.log(err);
-    res.status(500).json({
-      message: 'Не вдалось зареєструватись',
-    });
-  }
-});
-
+app.post('/auth/register', registerValidation, UserController.register);
 // Перевірка що ми можемо отримати інформацію про себе
-app.get('/auth/me', checkAuth, async (req, res) => {
-  try {
-    const user = await UserModel.findById(req.userId);
-
-    if (!user) {
-      return res.status(404).json({
-        message: 'Користувач не знайдений',
-      });
-    }
-    const { passwordHash, ...userData } = user._doc;
-
-    res.json(userData);
-  } catch (err) {
-    console.log(err);
-    res.status(500).json({
-      message: 'Нема доступу',
-    });
-  }
-});
+app.get('/auth/me', checkAuth, UserController.getMe);
 
 // порт серверу
 app.listen(4444, (err) => {
